Move workspace-related files to src/rmplan/workspace/ subdirectory
- Created new workspace subdirectory under src/rmplan/
- Moved all workspace_*.ts files to the new directory
- Updated all import paths to reflect the new file structure
- Fixed relative imports in moved files to account for deeper nesting
- Updated imports in agent.ts and rmplan.ts to reference the new paths
- All tests pass except for one pre-existing test failure
