+++ src/rmplan/workspace/workspace_manager.ts
@@ -1,9 +1,9 @@
 import * as fs from 'node:fs/promises';
 import * as path from 'node:path';
-import { debugLog, log } from '../logging.js';
-import { spawnAndLogOutput } from '../rmfilter/utils.js';
-import { executePostApplyCommand } from './actions.js';
-import type { PostApplyCommand, RmplanConfig } from './configSchema.js';
+import { debugLog, log } from '../../logging.js';
+import { spawnAndLogOutput } from '../../rmfilter/utils.js';
+import { executePostApplyCommand } from '../actions.js';
+import type { PostApplyCommand, RmplanConfig } from '../configSchema.js';
 import { WorkspaceLock } from './workspace_lock.js';
 import { getDefaultTrackingFilePath, recordWorkspace } from './workspace_tracker.js';
