+++ src/rmplan/workspace/workspace_tracker.ts
@@ -1,7 +1,7 @@
 import * as path from 'node:path';
 import * as os from 'node:os';
 import * as fs from 'node:fs/promises';
-import { log } from '../logging.js';
+import { log } from '../../logging.js';
 import { WorkspaceLock, type LockInfo } from './workspace_lock.js';
 
 /**
